Cleaned up code for project submission
